Better support for checking for a subset of errors.
Changed DiagnosticsReport.get_errors to DiagnosticsReport.has_errors
which accepts an optional list/set of keys to check for errors. This
allows the calling code to quickly determine if a subset of errors is
present in the report.

For example in manufacturing we only are interested in a subset of
errors so this makes it easier to call from there instead of getting all
errors and then doing such intersections there.